refactor: remove unused file and split handler
